create and list transfers server side

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/TransferDao.java
@@ -14,7 +14,7 @@ public interface TransferDao{
 
      Transfers pendingTransactions();
 
-     Transfers findTransfer(int id) throws Exception;
+     Transfers findTransfer(long id) throws Exception;
 
      Transfers createTransfer(Transfers transfers);
